ReactivePropertyResolverに統合
