+++ src/MessagePack.ReactiveProperty/MessagePack.ReactiveProperty.csproj
@@ -13,6 +13,11 @@
   </PropertyGroup>
 
   <ItemGroup>
+    <Compile Remove="ReactivePropertySlimResolver.cs" />
+    <Compile Remove="SlimFormatter.cs" />
+  </ItemGroup>
+
+  <ItemGroup>
     <PackageReference Include="System.Reactive.PlatformServices" Version="3.1.1" />
   </ItemGroup>
